// src/app/model/Category.model.ts
// export class Category{

//     id?:number | undefined;
//     categoryCode: number | undefined;
//     categoryName: string | undefined;
//     description: string | undefined;
// }

export class Category {
  id?: number | undefined;
  // categoryCode: number | undefined;
  categoryName: string | undefined;
  showSub?: boolean;
}