// src/app/model/Product.model.ts
export class Product {
  id?: number | undefined;
  productCode: string | undefined;
  productName: string | undefined;
  description: string | undefined;
  categoryId: number | undefined;
  categoryName: string | undefined;
  subcategoryId: number | undefined;
  subcategoryName: string | undefined;
  sellerId: number | undefined;
  price: number | undefined;
  stock: number | undefined;
  mainImage: string | undefined;
}